requirements.txt
autopep8==2.0.2
pycodestyle==2.10.0
python-dotenv==1.0.0
tomli==2.0.1
